requirements.txt
astroid==2.2.5
autopep8==1.4.4
backcall==0.1.0
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.0
dj-database-url==0.5.0
Django==2.2.2
django-bootstrap4==0.0.8
django-cors-headers==2.5.2
django-extensions==2.1.6
djangorestframework==3.9.2
drf-yasg==1.15.0
google-api-core==1.9.0
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-language==1.2.0
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.9
grpcio==1.19.0
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
ipython==7.4.0
ipython-genutils==0.2.0
isort==4.3.20
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
parso==0.4.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.5.0
Pygments==2.3.1
pylint==2.3.1
pytz==2018.9
requests==2.21.0
rsa==4.0
ruamel.yaml==0.15.92
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
whitenoise==4.1.2
wrapt==1.11.2
